Move button enable and toggle functions into image sync mixin
Refactor as static methods within the mixin for encapsulation.
